+++ kmath-core/src/commonMain/kotlin/space/kscience/kmath/nd/Structure1D.kt
@@ -117,12 +117,13 @@ internal class MutableBuffer1DWrapper<T>(val buffer: MutableBuffer<T>) : Mutable
 /**
  * Represent a [StructureND] as [Structure1D]. Throw error in case of dimension mismatch.
  */
-public fun <T> StructureND<T>.as1D(): Structure1D<T> = this as? Structure1D<T> ?: if (shape.size == 1) {
-    when (this) {
-        is BufferND -> Buffer1DWrapper(this.buffer)
-        else -> Structure1DWrapper(this)
-    }
-} else error("Can't create 1d-structure from ${shape.size}d-structure")
+public fun <T> StructureND<T>.as1D(): Structure1D<T> =
+    this as? Structure1D<T> ?: if (shape.size == 1) {
+        when (this) {
+            is BufferND -> Buffer1DWrapper(this.buffer)
+            else -> Structure1DWrapper(this)
+        }
+    } else error("Can't create 1d-structure from ${shape.size}d-structure")
 
 public fun <T> MutableStructureND<T>.as1D(): MutableStructure1D<T> =
     this as? MutableStructure1D<T> ?: if (shape.size == 1) {
@@ -133,13 +134,3 @@ public fun <T> MutableStructureND<T>.as1D(): MutableStructure1D<T> =
  * Represent this buffer as 1D structure
  */
 public fun <T> Buffer<T>.asND(): Structure1D<T> = Buffer1DWrapper(this)
-
-/**
- * Expose inner buffer of this [Structure1D] if possible
- */
-internal fun <T : Any> Structure1D<T>.asND(): Buffer<T> = when {
-    this is Buffer1DWrapper<T> -> buffer
-    this is Structure1DWrapper && structure is BufferND<T> -> structure.buffer
-    else -> this
-}
-